Enable visual bells on GNU Screen.
PuTTY has a feature that blinks its taskbar button when a bell occurs
and the window is not in focus. This feature is useful to notify the
termination of long-running commands. In order for this feature to work
even for bells occuring on the GNU Screen, `vbell` should be set to
`on`.