Revert aa19f078f7097159ec2fdbbc592dca371e5696dd.
PuTTY has a feature that blinks its taskbar button when a bell occurs
and the window is not in focus. This feature is useful to notify the
termination of long-running commands. In order for this feature to work
even for bells occuring on the GNU Screen, `vbell` should be set to
`off`.

By the commit aa19f078f7097159ec2fdbbc592dca371e5696dd, `vbell` is
wrongly set to `on`. This commit reverts it.