Add support for `gnu_screen_helper` to be installed to system path
The `gnu_screen_helper` script was originally designed to be installed
under `~/.local`.  Although the dependencies of `gnu_screen_helper`
installation include binary files specific to each environment, there
was no problem so far because the home directory was not assumed to be
shared by multiple environments.  However, because the home directory
is now shared among multiple environments, especially ones that are
created as Docker containers, the original design no longer works.
Therefore, it became necessary to allow `gnu_screen_helper` to be
installed to an environment-specific location such as `/usr/local`.
This commit makes this possible.